temporary fix for coordinate transform
the computed gmst was actually the local mean sidereal time, due to the fact that we passed the local time. If we want to still name the function Gmst, then we would only be allowed to pass the relative utc time. For now we leave it as Lmst, but it could produce some issues in the future.